+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # CHANGELOG
 
-*1.2.0* (2021-11-17)
+*Unreleased*
+- Set group for permissions from environnement variables
 - Set user notifications on error persistant
-
-*1.1.0* (2021-11-17)
 - Rework User information, so the admin can see the current version
 
 *1.0.0* (2021-11-17)